requirements.txt
streamlit==1.41.1
python-dotenv==1.0.1
openai==1.57.4
requests==2.32.3

# Additional dependencies
passlib
requests
oauthlib
requests-oauthlib
sqlite3